libcfnet/bwlimit: code style improvements, types, comments
Several points, after the code revision. Still, code will act exactly
as before this commit.

Edward Welbourne wrote:
> Style: open brace on its own line, not at end of previous line.
>
> Indentation and line-splitting is odd.  Please put the && at the end
> of the previous line, so the next line's bwlimit_next lines up under
> that of the previous line, making it easier to see what they have in
> common and where they differ.  Also, the part after || should be lined
> up under the part before; if it looks to you like it does, check for
> tabs (please use spaces for indent); it doesn't appear lined up in the
> code-review !
>
> This 1000000000L is kinda hard to read.  ...
>
> next and lock can (so should) be static; they're only referenced from
> this file.

Dimitris Apostolou wrote:
> maybe you can replace the u_long type with C89/C99 type?

+++ libpromises/generic_agent.h
@@ -126,5 +126,5 @@ bool GenericAgentTagReleaseDirectory(const GenericAgentConfig *config, const cha
 
 void GetReleaseIdFile(const char *base_path, char *filename, size_t max_size);
 
-extern u_long bwlimit_kbytes; /* from libcfnet/net.c */
+extern uint32_t bwlimit_kbytes; /* from libcfnet/net.c */
 #endif
